backend/requirements.txt
fastapi
uvicorn
sqlalchemy
openai==0.28.0
numpy
pydantic
python-dotenv